+++ apps/portal/src/app/api/bff/auth/github/callback/route.test.ts
@@ -25,7 +25,7 @@ vi.mock("next/headers", () => ({
 }));
 
 vi.mock("@portal/server/bff/backend", () => ({
-  deploymentClient: vi.fn(async () => ({
+  backendClient: vi.fn(async () => ({
     exchangeGitHubCode: mocks.exchangeGitHubCode,
   })),
 }));
