+++ packages/deploy/src/backend/index.ts
@@ -1,0 +1,2 @@
+export { BackendClient } from "./user";
+export { assertServerOnly } from "./core";
