+++ packages/deploy/src/wire/index.ts
@@ -1,0 +1,3 @@
+export * from "./requests";
+export * from "./responses";
+export * from "./operate";
